backend/requirements.txt
# Web 框架
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.9

# LLM
langchain==0.3.19
langchain-core==0.3.45
langchain-anthropic>=0.3.0

# 工具
python-dotenv==1.1.1
httpx==0.27.2

# 数据处理
pandas==2.2.3

# 数据库
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Google Maps
googlemaps==4.10.0

# 其他
python-multipart==0.0.9

